数组实现

+++ src/main/java/runtime/rtda/priv/OperandStack.java
@@ -25,7 +25,7 @@ public class OperandStack {
         pushSlot(slot);
     }
 
-    public int popByte() {
+    public byte popByte() {
         return (byte) popSlot().get();
     }
